Implemented tab bar for main home user page
