+++ lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart'
     show BuildContext, MaterialApp, StatelessWidget, Widget, runApp;
 import 'package:winhacks/screens/create_events.dart';
+import 'package:winhacks/screens/home/home.dart';
 import 'package:winhacks/screens/log_in.dart';
 import 'package:winhacks/screens/navigation_bar.dart';
 import 'package:winhacks/screens/wrapper.dart';
@@ -19,9 +20,9 @@ class MyApp extends StatelessWidget {
       value: AuthService().user,
       initialData: null,
       child: MaterialApp(
-        //home: Wrapper(),
-        home: Navigation_Bar(),
-      ),
+          //home: Wrapper(),
+          home: Navigation_Bar() //Navigation_Bar(),
+          ),
     );
   }
 }
